# api.yaml
...
openapi: 3.1.0
info:
  title: Moodle Replica API
  description: Skeleton API for the people & enrolment module.
  version: 0.1.0
paths:
...
  /api/users/{user_id}:
    get:
      tags:
      - users
      summary: Get User
      operationId: get_user_api_users__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      - name: x-acting-user
        in: header
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: X-Acting-User
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/courses:
    get:
      tags:
      - courses
      summary: List Courses
      operationId: list_courses_api_courses_get
      parameters:
      - name: include_deleted
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Include Deleted
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/courses/{course_id}:
    get:
      tags:
      - courses
      summary: Get Course
      operationId: get_course_api_courses__course_id__get
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer